aqrosmart/backend/requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
sqlalchemy[asyncio]==2.0.20
alembic==1.12.0
asyncpg==0.28.0
pydantic-settings==2.0.3
psycopg2-binary==2.9.7
